Rename the dns plugin to not conflict with dns namespace
There is  python dns module and this plugin naming conflicts
with that.
